Nexa Chat Assistant/requirements.txt
Flask
gTTS
playsound
SpeechRecognition
yt_dlp
openai
python-dotenv
